requirements-poc.txt
openai>=1.52.0
pydantic>=2.0.0
python-dotenv>=1.0.0
